Create update components, correct register to also be used by update
